requirements.txt
absl_py==2.3.1
aiofiles==24.1.0
attrs==25.4.0
bibtexparser==1.4.3
certifi==2025.10.5
charset_normalizer==3.4.3
chex==0.1.90
click==8.3.0
contourpy==1.3.1
cycler==0.12.1
dm-haiku @ git+https://github.com/deepmind/dm-haiku@c7b81923dbf20bbc662dc961fc2a33d1c3645703
e3nn-jax @ git+https://github.com/e3nn/e3nn-jax.git@403f0ef159c537c2efa9d3d05799da85a86575d5
einops==0.8.1
equinox==0.13.0
etils==1.13.0
filelock==3.20.0
flax==0.11.2
fonttools==4.60.0
freud_analysis==3.3.1
frozendict==2.4.6
fsspec==2025.9.0
GridDataFormats==1.0.2
grpcio==1.73.0
humanize==4.13.0
idna==3.10
importlib_resources==6.5.2
jax==0.6.0
jax-md==0.2.27
jaxlib==0.6.0
jaxopt==0.8.5
jaxtyping==0.3.2
jmp==0.0.4
joblib==1.5.2
jraph==0.0.6.dev0
jsonschema==4.25.1
jsonschema_specifications==2025.9.1
kiwisolver==1.4.8
lineax==0.0.8
lipyphilic @ https://github.com/p-j-smith/lipyphilic/archive/main.zip#sha256=2faef30298578d08e8c6e1f679d36e13ef2350e3679e0205307586437536da9c
markdown_it_py==4.0.0
matplotlib==3.10.0
mda_xdrlib==0.2.0
MDAnalysis==2.9.0
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.1
mmtf_python==1.1.3
monty==2025.3.3
more_itertools==10.8.0
mpmath==1.3.0
mrcfile==1.5.3
msgpack==1.1.0
narwhals==2.5.0
nest_asyncio==1.6.0
networkx==3.5
numpy==2.2.2
opt_einsum==3.4.0
optax==0.2.6
optimistix @ git+https://github.com/patrick-kidger/optimistix@396aac6fa66f6b6f1b6e3d3538cb77003362979a
orbax_checkpoint==0.11.25
ott-jax @ git+https://github.com/ott-jax/ott@458e265a1282d417f87bc0e07a5b6b77ffe9ae50
packaging==25.0
palettable==3.3.3
pandas==2.2.3
parsnip_cif==0.2.1
pillow==11.1.0
plotly==6.3.0
protobuf==6.32.1
pygments==2.19.2
pymatgen==2025.4.19
pyparsing==3.2.5
python_dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ray==2.49.1
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rowan==1.3.2
rpds_py==0.27.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scipy==1.15.1
seaborn==0.13.2
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
spglib==2.5.0
sympy==1.14.0
tabulate==0.9.0
tensorstore==0.1.76
threadpoolctl==3.6.0
tidynamics @ git+https://github.com/pdebuyl-lab/tidynamics.git@b0c95e0b337ea1c9748bfebf1ac710182360ef53
toolz==1.0.0
tqdm==4.67.1
treescope==0.1.10
typing_extensions==4.15.0
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.5.0
wadler_lindig==0.1.7
zipp==3.23.0
